+++ test/functional/initialization.bats.sh
@@ -11,7 +11,7 @@ load test_helper
   run ls $BATS_TEST_DIRNAME/.ttt.ini
   [ "$status" -eq 1 ]
   # Assert running the program creates ini
-  run $BATS_TEST_DIRNAME/tsp
+  run $BATS_TEST_DIRNAME/ttt
   run ls $BATS_TEST_DIRNAME/.ttt.ini
   [ "$status" -eq 0 ]
 }
@@ -21,14 +21,14 @@ load test_helper
   run ls $BATS_TEST_DIRNAME/timesheet.html
   [ "$status" -eq 1 ]
   # Assert running the program creates timesheet
-  run $BATS_TEST_DIRNAME/tsp
+  run $BATS_TEST_DIRNAME/ttt
   run ls $BATS_TEST_DIRNAME/timesheet.html
   [ "$status" -eq 0 ]
 }
 
 @test "The report-path in a default ini is the path of the executable" {
   # Create ini
-  run $BATS_TEST_DIRNAME/tsp
+  run $BATS_TEST_DIRNAME/ttt
   # Compare report-path
   run grep ";report_path=" $BATS_TEST_DIRNAME/.ttt.ini
   [[ "$output" = *$BATS_TEST_DIRNAME* ]]