+++ test/functional/merge.bats.sh
@@ -7,10 +7,10 @@
 load test_helper
 
 @test 'If there are no entries, "merge" fails w/ error message' {
-  run $BATS_TEST_DIRNAME/tsp
+  run $BATS_TEST_DIRNAME/ttt
 
   # Merge fails
-  $BATS_TEST_DIRNAME/tsp m 1 | grep 'Cannot merge'
+  $BATS_TEST_DIRNAME/ttt m 1 | grep 'Cannot merge'
 
   # There are no entries
   run grep -c '<td class="meta">/' $BATS_TEST_DIRNAME/timesheet.html
@@ -18,106 +18,106 @@ load test_helper
 }
 
 @test 'After "merge" there is one entry less than before' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp p
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt p
   # There are 2 entries
   run grep -c '<td class="meta">' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 2 ]]
 
-  run $BATS_TEST_DIRNAME/tsp m 0
+  run $BATS_TEST_DIRNAME/ttt m 0
   # There is 1 entry
   run grep -c '<td class="meta">' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test '"merge" of two entries with a break inbetween fails' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp s i=0 12:30
-  run $BATS_TEST_DIRNAME/tsp p i=0 12:35
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp s i=1 12:40
-  run $BATS_TEST_DIRNAME/tsp p i=1 12:45
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt s i=0 12:30
+  run $BATS_TEST_DIRNAME/ttt p i=0 12:35
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt s i=1 12:40
+  run $BATS_TEST_DIRNAME/ttt p i=1 12:45
   # Merge fails
-  $BATS_TEST_DIRNAME/tsp m 0 | grep 'Cannot merge: Entries have a gap of 5 minutes'
+  $BATS_TEST_DIRNAME/ttt m 0 | grep 'Cannot merge: Entries have a gap of 5 minutes'
 }
 
 @test '"merge" of the last entry fails w/ an error message' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp p
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt p
   # Merge fails
-  $BATS_TEST_DIRNAME/tsp m 0 | grep 'Cannot merge'
+  $BATS_TEST_DIRNAME/ttt m 0 | grep 'Cannot merge'
 }
 
 @test 'Merging two entries retains the comment of the 1st entry if the other has none' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m 1
   run grep -c 'foo' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging two entries retains the comment of the 2nd entry if the other has none' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m 1
   run grep -c 'foo' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging two entries retains the task-number of the 1st entry if the other has none' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s 123 foo
-  run $BATS_TEST_DIRNAME/tsp s bar
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s 123 foo
+  run $BATS_TEST_DIRNAME/ttt s bar
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m 1
   run grep -c '123' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging two entries retains the task-number of the 2nd entry if the other has none' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp s 123 bar
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt s 123 bar
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m 1
   run grep -c '123' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging two entries that have different comments, appends them' {
-  run $BATS_TEST_DIRNAME/tsp s "fixed bug"
-  run $BATS_TEST_DIRNAME/tsp s "tested"
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m
+  run $BATS_TEST_DIRNAME/ttt s "fixed bug"
+  run $BATS_TEST_DIRNAME/ttt s "tested"
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m
   run grep -c 'fixed bug. tested' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging two entries that have the same comment, retains it once' {
-  run $BATS_TEST_DIRNAME/tsp s
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp p
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt s
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt p
+  run $BATS_TEST_DIRNAME/ttt m 1
   run grep -c 'foo' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
 }
 
 @test 'Merging a stopped w/ a running entry, sets the merged entry running' {
-  run $BATS_TEST_DIRNAME/tsp s foo
-  run $BATS_TEST_DIRNAME/tsp s
+  run $BATS_TEST_DIRNAME/ttt s foo
+  run $BATS_TEST_DIRNAME/ttt s
   # There is 1 running entry
   run grep -c '<td class="meta">s/' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]
   # Merge
-  run $BATS_TEST_DIRNAME/tsp m 1
+  run $BATS_TEST_DIRNAME/ttt m 1
   # There is 1 running entry
   run grep -c '<td class="meta">s' $BATS_TEST_DIRNAME/timesheet.html
   [[ "$output" = 1 ]]